+++ app/src/main/java/cityfreqs/com/pilfershushjammer/AudioSettings.java
@@ -37,7 +37,7 @@ class AudioSettings {
             "audioSource", "sampleRate", "channelInConfig", "encoding",
             "bufferInSize", "channelOutConfig", "bufferOutSize", "activeType",
             "jammerType", "userCarrier", "userLimit", "userSpeed", "hasEQ",
-            "maxFreq"
+            "maxFreq", "bufferRead"
     };
 }