+++ app/src/main/java/cityfreqs/com/pilfershushjammer/PassiveJammer.java
@@ -8,10 +8,12 @@ public class PassiveJammer {
     private Context context;
     private Bundle audioBundle;
     private AudioRecord audioRecord;
+    private boolean READ_BUFFER;
 
     public PassiveJammer(Context context, Bundle audioBundle) {
         this.context = context;
         this.audioBundle = audioBundle;
+        READ_BUFFER = false;
     }
 
     boolean startPassiveJammer() {
@@ -49,21 +51,25 @@ public class PassiveJammer {
                     audioRecord.startRecording();
                     MainActivity.entryLogger(context.getResources().getString(R.string.passive_state_3) + "\n", true);
 
-                    // check for initialising audioRecord
-                    short buffer[] = new short[audioBundle.getInt(AudioSettings.AUDIO_BUNDLE_KEYS[4])];
+                    // optional switch for accessing the hardware buffers via audioRecord.read()
+                    // reason: possible battery use at hardware level
+                    if (READ_BUFFER) {
+                        // check for initialising audioRecord
+                        short buffer[] = new short[audioBundle.getInt(AudioSettings.AUDIO_BUNDLE_KEYS[4])];
 
-                    // returns either 0, number of shorts read, or an error code - not audio data
-                    int audioStatus = audioRecord.read(buffer, 0,  audioBundle.getInt(AudioSettings.AUDIO_BUNDLE_KEYS[4]));
-                    // check for error on pre 6.x and 6.x API
-                    if (audioStatus == AudioRecord.ERROR_INVALID_OPERATION) {
-                        MainActivity.entryLogger(context.getResources().getString(R.string.passive_state_4), true);
-                        // error from improper use of method
-                        return false;
-                    } else if (audioStatus == AudioRecord.STATE_UNINITIALIZED) {
-                        MainActivity.entryLogger(context.getResources().getString(R.string.passive_state_5_1), true);
-                        MainActivity.entryLogger(context.getResources().getString(R.string.passive_state_5_2), true);
-                        // adb:  status code -38
-                        return false;
+                        // returns either 0, number of shorts read, or an error code - not audio data
+                        int audioStatus = audioRecord.read(buffer, 0, audioBundle.getInt(AudioSettings.AUDIO_BUNDLE_KEYS[4]));
+                        // check for error on pre 6.x and 6.x API
+                        if (audioStatus == AudioRecord.ERROR_INVALID_OPERATION) {
+                            MainActivity.entryLogger(context.getResources().getString(R.string.passive_state_4), true);
+                            // error from improper use of method
+                            return false;
+                        } else if (audioStatus == AudioRecord.STATE_UNINITIALIZED) {
+                            MainActivity.entryLogger(context.getResources().getString(R.string.passive_state_5_1), true);
+                            MainActivity.entryLogger(context.getResources().getString(R.string.passive_state_5_2), true);
+                            // adb:  status code -38
+                            return false;
+                        }
                     }
 
                     // check for running audioRecord